# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Booking API",
    "description": "API для системи бронювання (користувачі та записи)",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-production.up.railway.app"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login a user",
        "operationId": "loginUser",
        "description": "Login a user with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "example": "Jonni_Do1@gmail.com"
          },
          {
            "in": "query",
            "name": "password",
            "schema": {
              "type": "string"
            },
            "example": "111223Jn"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "description": "User's email",
                    "type": "string",
                    "example": "John_Doe123@gmail.com"
                  },
                  "password": {
                    "description": "password",
                    "type": "string",
                    "example": "111223Jn"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully login a user!"
                    },
                    "data": {
                      "type": "object",
                      "required": ["accessToken"],
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "d9H4sXl8AqP5oC2mJZvTNlGuK03xM1r"
                        },
                        "refreshToken": {
                          "type": "string",
                          "example": "d9H4sXl8AqP5oC2mJZvTNlGuK03xM1rLtWOKFbY7"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a booking",
        "operationId": "createBooking",
        "description": "Create a booking for a business user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["businessId", "date", "time"],
                "properties": {
                  "businessId": {
                    "type": "string",
                    "description": "ID бізнес-користувача, у якого робимо бронь",
                    "example": "64adf1c2a1b23cd456ef890a"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "Дата бронювання",
                    "example": "2025-09-10"
                  },
                  "time": {
                    "type": "string",
                    "description": "Час бронювання",
                    "example": "14:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "clientId",
                    "businessId",
                    "date",
                    "time",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "64adf1c2a1b23cd456ef890a"
                    },
                    "clientId": {
                      "type": "string",
                      "description": "ID клієнта",
                      "example": "64adf1c2a1b23cd456ef1111"
                    },
                    "businessId": {
                      "type": "string",
                      "description": "ID бізнес-користувача",
                      "example": "64adf1c2a1b23cd456ef2222"
                    },
                    "date": {
                      "type": "string",
                      "format": "date",
                      "example": "2025-09-10"
                    },
                    "time": {
                      "type": "string",
                      "example": "14:00"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["pending", "confirmed", "cancelled"],
                      "example": "pending"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BadRequestError",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "BadRequestError"
                    },
                    "data": {
                      "type": "object",
                      "required": ["message"],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad request"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1contacts/get/responses/404"
          }
        }
      }
    },
    "/bookings/me": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get my bookings",
        "operationId": "getMyBookings",
        "description": "Get all bookings for the authenticated client",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/paths/~1bookings/post/responses/201/content/application~1json/schema"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          }
        }
      }
    },
    "/bookings/{id}": {
      "delete": {
        "tags": ["Bookings"],
        "summary": "Delete booking",
        "operationId": "deleteBooking",
        "description": "Delete a booking by ID (only the owner can delete)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Booking ID",
              "example": "64adf1c2a1b23cd456ef890a"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Booking successfully deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1contacts/get/responses/404"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/bookings/{id}/cancel": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Cancel booking",
        "operationId": "cancelBooking",
        "description": "Cancel a booking by ID (status changes to cancelled)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "64adf1c2a1b23cd456ef890a"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking successfully cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1bookings/post/responses/201/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1bookings/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1contacts/get/responses/404"
          }
        }
      }
    },
    "/bookings/{id}/update": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Update booking",
        "operationId": "updateBooking",
        "description": "Update booking details (time, contact info, number, etc.)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-09-10T14:00:00Z"
                  },
                  "contact": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+380501234567"
                  },
                  "tableNumber": {
                    "type": "integer",
                    "example": 12
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Booking updated successfully"
                    },
                    "data": {
                      "$ref": "#/paths/~1bookings/post/responses/201/content/application~1json/schema"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1bookings/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1contacts/get/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1contacts/get/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}